--- src/Abp.EntityFrameworkCore/EntityFrameworkCore/IDbContextTypeMatcher.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace Abp.EntityFrameworkCore
-{
-    public interface IDbContextTypeMatcher
-    {
-        void Populate(Type[] dbContextTypes);
-
-        Type GetConcreteType(Type sourceDbContextType);
-    }
-}
